requirements.txt
pip==23.1.2
psycopg2-binary==2.9.1